refactor(dapp): split out executor
